+++ src/operations/queries/getPaginatedOrganizationUsers.tsx
@@ -8,7 +8,7 @@ import {
 } from "@/types/graphQL";
 import { isUuid } from "@/utils/pagination";
 import { gql } from "@apollo/client";
-import { BaseTableData } from "@kl-engineering/kidsloop-px/dist/types/components/Table/Common/BaseTable";
+import { BaseTableData } from "@kl-engineering/kidsloop-px/dist/src/components/Table/Common/BaseTable";
 
 export interface ProgramQueryFilter {
     organizationId: string;