+++ src/pages/admin/roles/index.tsx
@@ -15,9 +15,9 @@ import {
     serverToTableOrder,
     tableToServerOrder,
 } from "@/utils/table";
-import { Order } from "@kl-engineering/kidsloop-px/dist/types/components/Table/Common/Head";
-import { PageChange } from "@kl-engineering/kidsloop-px/dist/types/components/Table/Common/Pagination/shared";
-import { CursorTableData } from "@kl-engineering/kidsloop-px/dist/types/components/Table/Cursor/Table";
+import { Order } from "@kl-engineering/kidsloop-px/dist/src/components/Table/Common/Head";
+import { PageChange } from "@kl-engineering/kidsloop-px/dist/src/components/Table/Common/Pagination/shared";
+import { CursorTableData } from "@kl-engineering/kidsloop-px/dist/src/components/Table/Cursor/Table";
 import React,
 {
     useEffect,
@@ -132,6 +132,7 @@ export default function RolesPage () {
             startCursor={pageInfo?.startCursor}
             endCursor={pageInfo?.endCursor}
             onPageChange={handlePageChange}
-            onTableChange={handleTableChange}/>
+            onTableChange={handleTableChange}
+        />
     );
 }